feature: jest 특강 관련 자료 업데이트 ... 불태웠다

+++ jest특강/분리후/user.service.js
@@ -4,32 +4,36 @@ export class UserService {
     }
 
     // 새 사용자 생성
-    addUser(user) {
-        // 유효성 검사 및 비즈니스 로직 처리
-        // 예: user 객체에 필수 필드가 있는지 확인
-        if (!user.id || !user.name) {
-            throw new Error('ID와 이름은 필수입니다.');
-        }
-        return this.userRepository.createUser(user);
+    createUser = async (id, name) => {
+        const user = { id, name };
+        return await this.userRepository.createUser(user);
     }
 
     // 모든 사용자 조회
-    getAllUsers() {
-        return this.userRepository.findAllUsers();
+    getAllUsers = async () => {
+        return await this.userRepository.findAllUsers();
     }
 
     // ID로 사용자 조회
-    getUserById(userId) {
-        return this.userRepository.findUserById(userId);
+    getUserById = async (userId) => {
+        return await this.userRepository.findUserById(userId);
     }
 
     // 사용자 정보 업데이트
-    updateUser(userId, userData) {
-        return this.userRepository.updateUser(userId, userData);
+    updateUser = async (userId, id, name) => {
+        const targetUser = await this.userRepository.findUserById(userId);
+        if (targetUser) {
+            return await this.userRepository.updateUser(userId, id, name);
+        }
+        return null;
     }
 
     // 사용자 삭제
-    removeUser(userId) {
-        return this.userRepository.deleteUser(userId);
+    removeUser = async (userId) => {
+        const targetUser = await this.userRepository.findUserById(userId);
+        if (targetUser) {
+            return await this.userRepository.deleteUser(userId);
+        }
+        return false;
     }
 }
